finish show tv show, now prepare login

+++ app/src/main/java/com/ferdsapp/moviefinder/core/data/model/network/nowPlaying/tvShow/ItemTvShowPlaying.kt
@@ -40,7 +40,7 @@ data class ItemTvShowPlaying(
     val name: String,
 
     @field:SerializedName("vote_average")
-    val vote_average: Int,
+    val vote_average: Float,
 
     @field:SerializedName("vote_count")
     val vote_count: Int
